requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==2.2.5
idna==3.10
importlib-metadata==6.7.0
itsdangerous==2.1.2
jinja2==3.1.4
MarkupSafe==2.1.5
python-dateutil==2.9.0.post0
requests==2.31.0
six==1.16.0
typing-extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
zipp==3.15.0
